+++ CHANGELOG.rst
@@ -3,6 +3,7 @@
 - Add ``receiver_connections.XXX.csv`` which summarizes connections, disconnections and slow receivers.
 - Improve heuristics for grouping related warn/error messages to include filtering of stringified events, and of text at least 80 characters through the string that's following a colon.
 - Add ``README.rst`` instructions for opening CSV files in non-English locales such as German. Unfortunately it's quite an involved process. 
+- Add ``--skip 10%`` command line option for skipping the beginning of each file to avoid startup noise, increase focus on the end (where is usually where the juicy bits are), and allow the tool to run faster. If the log begins with startup ``#####`` messages these will still be read before skipping forwards. 
 
 3.1
 ---